Fix status parsing of file:/ feeds on Python>=3.9
Starting with feedparser 6.0.0b1, the return value from feedparser.parse()
on a file:/ feed always has the status attribute. However, it attempts to set
the attribute to 200 where it was previously absent, but this attempt breaks on
Python 3.9, and a None value slips through.

This inconsistency between feed types and between Python versions may be fixed
in a later release of feedparser, but in the meantime we can support the
already-released 6.0.x versions.